Move MOverflowRaw from vhf to vhf_common
Centralize between its use in process/fold and in parsing.
Consequently, MOverflowWrite has to properly change to a NewType
wrapper. To make that intent clear, it has been annotated with
`#[repr(transparent)] `.
