+++ VHF-rs/src/runner/process/test_vhf_step_fold.rs
@@ -4,11 +4,11 @@
 //! ```
 //! behaves to expectation.
 
-use super::super::fold::{MOverflowRaw, StreamFold, StreamFoldOp};
+use super::super::fold::{StreamFold, StreamFoldOp};
 use super::consts::MMAP_PAGE_LEN;
 use super::test_vhf::{debug_vhf_new, push_arc_pages};
 use super::*;
-use vhf_common::data_types::{IQMTriplet, Polar, RawVHFWord};
+use vhf_common::data_types::{IQMTriplet, MOverflowRaw, Polar, RawVHFWord};
 
 use std::f64::consts::{PI, TAU};
 use std::sync::atomic::{AtomicUsize, Ordering};
